fix returning cdn urls
