+++ assetman/tools.py
@@ -2,13 +2,27 @@ from __future__ import absolute_import, with_statement
 
 import os
 import re
+import binascii
 import itertools 
 
 # What to calls to assetman look like in {% apply %} blocks?
 include_expr_matcher = re.compile(r'^assetman\.(include_\w+)').match
+def get_shard_from_list(settings_list, shard_id):
+    assert isinstance(settings_list, (list, tuple)), "must be a list not %r" % settings_list
+    shard_id = _crc(shard_id)
+    bucket = shard_id % len(settings_list)
+    return settings_list[bucket]
 
-def _utf8(instr):
-    return instr.encode("utf-8")
+def _crc(key):
+    """crc32 hash a string"""
+    return binascii.crc32(_utf8(key)) & 0xffffffff
+
+def _utf8(s):
+    """encode a unicode string as utf-8"""
+    if isinstance(s, unicode):
+        return s.encode("utf-8")
+    assert isinstance(s, str), "_utf8 expected a str, not %r" % type(s)
+    return s
 
 def iter_template_paths(template_dirs, template_ext):
     """Walks each directory in the given list of template directories,
